Module questionnaire_rs::models

Structs

Answer

AnswerInput
The structure of the body of JSON request for creating a new answer.

GHAccessTokenBody
The structure of the body of JSON request for retrieving the access token for a session code.

GetAnswerById
This defines an actor for retrieving answer from database by id.

GetAnswersByOption
Defines an actor for retrieving answers for a question.

GetOption
Defines an actor to retrieve an option from database by id.

GetOptionsByQuestion
Defines an actor for retrieving options for a question.

GetPresentation
Defines an actor to retrieve presentation from database by id.

GetQuestion
This defines an actor for retrieving question from database by id.

GetQuestionByPresentation
Defines an actor for retrieving questions for a presentation.

NewAnswer
Creates a new answer.

NewOption
Creates a new option.

NewOptionJson
Defines the structure of the body of JSON request for creating a new option.

NewPresentation
Creates a new presentation.

NewQuestion
Creates a new question.

NewQuestionJson
The structure of the body of JSON request for creating a new question.

Option
Option model.

Presentation
Presentation model.

PresentationInput
The structure of the body of JSON request for creating a new presentation.

Questions

User